+++ src/main/java/com/project/btl/repository/CategoryRepository.java
@@ -1,3 +1,4 @@
+// File: com/project/btl/repository/CategoryRepository.java
 package com.project.btl.repository;
 
 import com.project.btl.model.entity.Category;
@@ -5,5 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Long> {
+// SỬA: Đổi JpaRepository<Category, Long> thành JpaRepository<Category, Integer>
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
 }
